refactor: replace shopify imports
