+++ app/product/[handle]/page.tsx
@@ -10,8 +10,8 @@ import { Gallery } from 'components/product/gallery';
 import { VariantSelector } from 'components/product/variant-selector';
 import Prose from 'components/prose';
 import { HIDDEN_PRODUCT_TAG } from 'lib/constants';
-import { getProduct } from 'lib/shopify';
-import { Image } from 'lib/shopify/types';
+import { getProduct } from 'lib/medusa';
+import { Image } from 'lib/medusa/types';
 
 export const runtime = 'edge';
