+++ app/search/[collection]/page.tsx
@@ -1,4 +1,4 @@
-import { getCollection, getCollectionProducts } from 'lib/shopify';
+import { getCollection, getCollectionProducts } from 'lib/medusa';
 import { Metadata } from 'next';
 import { notFound } from 'next/navigation';
