+++ components/cart/index.tsx
@@ -1,4 +1,4 @@
-import { createCart, getCart } from 'lib/shopify';
+import { createCart, getCart } from 'lib/medusa';
 import { cookies } from 'next/headers';
 import CartButton from './button';
