+++ components/grid/three-items.tsx
@@ -1,6 +1,6 @@
 import { GridTileImage } from 'components/grid/tile';
-import { getCollectionProducts } from 'lib/shopify';
-import type { Product } from 'lib/shopify/types';
+import { getCollectionProducts } from 'lib/medusa';
+import type { Product } from 'lib/medusa/types';
 import Link from 'next/link';
 
 function ThreeItemGridItem({
@@ -41,13 +41,13 @@ export async function ThreeItemGrid() {
 
   if (!homepageItems[0] || !homepageItems[1] || !homepageItems[2]) return null;
 
-  // const [firstProduct, secondProduct, thirdProduct] = homepageItems;
+  const [firstProduct, secondProduct, thirdProduct] = homepageItems;
 
   return (
     <section className="lg:grid lg:grid-cols-6 lg:grid-rows-2" data-testid="homepage-products">
-      {/* <ThreeItemGridItem size="full" item={firstProduct} background="purple" />
+      <ThreeItemGridItem size="full" item={firstProduct} background="purple" />
       <ThreeItemGridItem size="half" item={secondProduct} background="black" />
-      <ThreeItemGridItem size="half" item={thirdProduct} background="pink" /> */}
+      <ThreeItemGridItem size="half" item={thirdProduct} background="pink" />
     </section>
   );
 }
