+++ components/layout/navbar/index.tsx
@@ -1,10 +1,7 @@
 import Link from 'next/link';
-import { Suspense } from 'react';
 
-import Cart from 'components/cart';
-import CartIcon from 'components/icons/cart';
 import LogoIcon from 'components/icons/logo';
-import { Menu } from 'lib/shopify/types';
+import { Menu } from 'lib/medusa/types';
 import MobileMenu from './mobile-menu';
 import Search from './search';
 
@@ -43,10 +40,10 @@ export default async function Navbar() {
       </div>
 
       <div className="flex w-1/3 justify-end">
-        <Suspense fallback={<CartIcon className="h-6" />}>
-          {/* @ts-expect-error Server Component */}
-          <Cart />
-        </Suspense>
+        {/* <Suspense fallback={<CartIcon className="h-6" />}> */}
+        {/* @ts-expect-error Server Component */}
+        {/*  <Cart /> */}
+        {/* </Suspense> */}
       </div>
     </nav>
   );
