+++ components/layout/search/collections.tsx
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
 import { Suspense } from 'react';
 
-import { getCollections } from 'lib/shopify';
+import { getCollections } from 'lib/medusa';
 import FilterList from './filter';
 
 async function CollectionList() {
